Implémentation finale idée Hayk

+++ src/main/java/insa/aubin/devisbatiment/modele/Point.java
@@ -4,29 +4,29 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
 public class Point implements Dessin {
-    private double x;
-    private double y;
+    private float x;
+    private float y;
     public static double RAYON_IN_DRAW = 5;
     private Color color = Color.BLACK;
     
-    public Point(double x, double y){
+    public Point(float x, float y){
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
